pipeline/requirements.txt
# Data Pipeline for Parent Map HK
# Python 3.11+

# HTTP & Web Scraping
requests>=2.31.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10

# Google Sheets
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Data & Schema
pydantic>=2.5.0
python-dotenv>=1.0.0

# Utilities
tenacity>=8.2.0
tqdm>=4.66.0
click>=8.1.0

# Optional LLM (OpenAI/Anthropic)
openai>=1.3.0
anthropic>=0.8.0

# Dev & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
